pyproject.toml
[tool.poetry]
name = "internal-employee-rag-chatbot"
version = "0.1.0"
description = ""
authors = ["siesto1elemento <rohitojha9720@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openparse = "^0.7.0"
faiss-cpu = "^1.11.0"
numpy = "^2.2.6"
python-dotenv = "^1.1.0"
fastapi = "^0.115.12"
asyncio = "^3.4.3"
uvicorn = "^0.34.3"
websockets = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
